update ida scripts
